refactor: signup as module
