+++ studio_app/rout_handlers/__init__.py
@@ -8,5 +8,6 @@ __all__ = [
     'edit_appointment',
     'confirm_appointment',
     'done_appointment',
-    'cancel_appointment'
+    'cancel_appointment',
+    'signup'
 ]
